refactor: extract services to separate module to resolve import errors
