+++ src/bands/bands.module.ts
@@ -1,13 +1,10 @@
 import { Module } from '@nestjs/common';
 import { BandsResolver } from './bands.resolver';
-import { BandsService } from './bands.service';
 import { HttpModule } from '@nestjs/axios';
-import { ArtistsModule } from '../artists/artists.module';
-import { GenresModule } from '../genres/genres.module';
+import { ServicesModule } from '../services/services.module';
 
 @Module({
-  imports: [HttpModule, ArtistsModule, GenresModule],
-  providers: [BandsResolver, BandsService],
-  exports: [BandsService],
+  imports: [HttpModule, ServicesModule],
+  providers: [BandsResolver],
 })
 export class BandsModule {}
